Moved ProjectInfoFinder to a project package
