+++ webhook-service/test/ch/datascience/webhookservice/hookcreation/HookCreatorSpec.scala
@@ -36,6 +36,7 @@ import ch.datascience.webhookservice.hookcreation.HookCreator.HookCreationResult
 import ch.datascience.webhookservice.hookcreation.ProjectHookCreator.ProjectHook
 import ch.datascience.webhookservice.hookcreation.ProjectHookVerifier.HookIdentifier
 import ch.datascience.webhookservice.model.{HookToken, ProjectInfo}
+import ch.datascience.webhookservice.project.ProjectInfoFinder
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.Matchers._
 import org.scalatest.WordSpec
